refactor: refactor unnecessary context
