--- containers/provider/RoomProvider/index.ts
@@ -1,5 +1,0 @@
-import { UseRoomContextType } from "./RoomProvider";
-import RoomContextProvider from "./RoomProvider";
-export * from "./type";
-export type { UseRoomContextType };
-export default RoomContextProvider;
