--- hooks/context/useRoom.tsx
@@ -1,9 +1,0 @@
-import { useContext } from "react";
-
-import RoomContext from "@/contexts/RoomContext";
-
-const useRoom = () => {
-  return useContext(RoomContext);
-};
-
-export default useRoom;
